requirements.txt
google-generativeai>=0.8.0
edge-tts>=6.1.0
huggingface_hub>=0.24.0
requests>=2.31.0
moviepy>=1.0.3
Pillow>=10.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0
tenacity>=8.2.0
python-telegram-bot>=20.7
pytest>=8.0.0
pytest-asyncio>=0.23.0
